fix hot challenge

+++ src/main/java/com/example/onedaypiece/wep/proceed/end/EndDto.java
@@ -14,7 +14,7 @@ public class EndDto {
     private String challengeTitle;
     private String challengeContent;
     private ChallengeCategory challengeCategory;
-    private Integer challengeProgress;
+    private Long challengeProgress;
     private String challengeImgUrl;
     private LocalDate challengeStart;
     private LocalDate challengeEnd;
