+++ src/main/java/com/example/onedaypiece/wep/proceed/proceed/ProceedDto.java
@@ -16,7 +16,7 @@ public class ProceedDto {
     private String challengeTitle;
     private String challengeContent;
     private ChallengeCategory challengeCategory;
-    private Integer challengeProgress;
+    private Long challengeProgress;
     private String challengeImgUrl;
     private LocalDate challengeStart;
     private LocalDate challengeEnd;
